debugged/refactored example tests
