--- client/example.test.js
@@ -1,57 +1,0 @@
-import React, {useState} from 'react';
-import {cleanup, fireEvent, render} from '@testing-library/react';
-const frisby = require('frisby');
-
-const CheckboxWithLabel = ({labelOn, labelOff}) => {
-  const [isChecked, setIsChecked] = useState(false);
-
-  const onChange = () => {
-    setIsChecked(!isChecked);
-  };
-
-  return (
-    <label>
-      <input type="checkbox" checked={isChecked} onChange={onChange} />
-      {isChecked ? labelOn : labelOff}
-    </label>
-  );
-};
-
-var sum = (a, b) => {
-  return a + b;
-};
-
-test('adds 1 + 2 to equal 3', () => {
-  expect(sum(1, 2)).toBe(3);
-});
-
-describe('Google', () => {
-  beforeAll(async () => {
-    await page.goto('https://google.com');
-  });
-
-  it('should be titled "Google"', async () => {
-    await expect(page.title()).resolves.toMatch('Google');
-  });
-});
-
-// Note: running cleanup afterEach is done automatically for you in @testing-library/react@9.0.0 or higher
-// unmount and cleanup DOM after the test is finished.
-// afterEach(cleanup);
-
-xit('CheckboxWithLabel changes the text after click', () => {
-  const {queryByLabelText, getByLabelText} = render(
-    <CheckboxWithLabel labelOn="On" labelOff="Off" />,
-  );
-
-  expect(queryByLabelText(/off/i)).toBeTruthy();
-
-  fireEvent.click(getByLabelText(/off/i));
-
-  expect(queryByLabelText(/on/i)).toBeTruthy();
-});
-
-it('should be a teapot', function () {
-  return frisby.get('http://httpbin.org/status/418')
-    .expect('status', 418);
-});
