+++ example_tests/rtl.test.js
@@ -1,0 +1,44 @@
+/**
+ * @jest-environment jsdom
+ */
+
+import React, {useState} from 'react';
+import {cleanup, fireEvent, render} from '@testing-library/react';
+
+
+const CheckboxWithLabel = ({labelOn, labelOff}) => {
+  const [isChecked, setIsChecked] = useState(false);
+
+  const onChange = () => {
+    setIsChecked(!isChecked);
+  };
+
+  return (
+    <label>
+      <input type="checkbox" checked={isChecked} onChange={onChange} />
+      {isChecked ? labelOn : labelOff}
+    </label>
+  );
+};
+
+// Note: running cleanup afterEach is done automatically for you in @testing-library/react@9.0.0 or higher
+// unmount and cleanup DOM after the test is finished.
+// afterEach(cleanup);
+
+describe('RTL test', () => {
+
+  it('CheckboxWithLabel changes the text after click', () => {
+    const {queryByLabelText, getByLabelText} = render(
+      <CheckboxWithLabel labelOn="On" labelOff="Off" />
+    );
+
+    expect(queryByLabelText(/off/i)).toBeTruthy();
+
+    fireEvent.click(getByLabelText(/off/i));
+
+    expect(queryByLabelText(/on/i)).toBeTruthy();
+  });
+
+});
+
+
